unify error capture across handlers
