--- packages/twenty-server/src/engine/core-modules/graphql/utils/should-capture-exception.util.ts
@@ -1,43 +1,0 @@
-import { HttpException } from '@nestjs/common';
-
-import {
-  BaseGraphQLError,
-  ErrorCode,
-} from 'src/engine/core-modules/graphql/utils/graphql-errors.util';
-
-export const graphQLErrorCodesToFilterOut = [
-  ErrorCode.GRAPHQL_VALIDATION_FAILED,
-  ErrorCode.UNAUTHENTICATED,
-  ErrorCode.FORBIDDEN,
-  ErrorCode.NOT_FOUND,
-  ErrorCode.METHOD_NOT_ALLOWED,
-  ErrorCode.TIMEOUT,
-  ErrorCode.CONFLICT,
-  ErrorCode.BAD_USER_INPUT,
-];
-
-export const shouldCaptureException = (
-  exception: Error,
-  statusCode?: number,
-): boolean => {
-  if (
-    exception instanceof BaseGraphQLError &&
-    graphQLErrorCodesToFilterOut.includes(exception?.extensions?.code)
-  ) {
-    return false;
-  }
-
-  if (
-    exception instanceof HttpException &&
-    exception.getStatus() >= 400 &&
-    exception.getStatus() < 500
-  ) {
-    return false;
-  }
-
-  if (statusCode && statusCode >= 400 && statusCode < 500) {
-    return false;
-  }
-
-  return true;
-};
