+++ packages/twenty-server/src/engine/core-modules/message-queue/message-queue.explorer.ts
@@ -5,21 +5,21 @@ import {
   ModuleRef,
   createContextId,
 } from '@nestjs/core';
-import { Module } from '@nestjs/core/injector/module';
 import { Injector } from '@nestjs/core/injector/injector';
 import { InstanceWrapper } from '@nestjs/core/injector/instance-wrapper';
+import { Module } from '@nestjs/core/injector/module';
 
-import { MessageQueueWorkerOptions } from 'src/engine/core-modules/message-queue/interfaces/message-queue-worker-options.interface';
 import {
   MessageQueueJob,
   MessageQueueJobData,
 } from 'src/engine/core-modules/message-queue/interfaces/message-queue-job.interface';
+import { MessageQueueWorkerOptions } from 'src/engine/core-modules/message-queue/interfaces/message-queue-worker-options.interface';
 
+import { ExceptionHandlerService } from 'src/engine/core-modules/exception-handler/exception-handler.service';
 import { MessageQueueMetadataAccessor } from 'src/engine/core-modules/message-queue/message-queue-metadata.accessor';
 import { MessageQueueService } from 'src/engine/core-modules/message-queue/services/message-queue.service';
 import { getQueueToken } from 'src/engine/core-modules/message-queue/utils/get-queue-token.util';
-import { ExceptionHandlerService } from 'src/engine/core-modules/exception-handler/exception-handler.service';
-import { shouldFilterException } from 'src/engine/utils/global-exception-handler.util';
+import { shouldCaptureException } from 'src/engine/utils/global-exception-handler.util';
 
 interface ProcessorGroup {
   instance: object;
@@ -205,7 +205,7 @@ export class MessageQueueExplorer implements OnModuleInit {
       try {
         await instance[processMethodName].call(instance, job.data);
       } catch (err) {
-        if (!shouldFilterException(err)) {
+        if (shouldCaptureException(err)) {
           this.exceptionHandlerService.captureExceptions([err]);
         }
         throw err;
