+++ packages/twenty-server/src/queue-worker/queue-worker.ts
@@ -2,7 +2,7 @@ import { NestFactory } from '@nestjs/core';
 
 import { ExceptionHandlerService } from 'src/engine/core-modules/exception-handler/exception-handler.service';
 import { LoggerService } from 'src/engine/core-modules/logger/logger.service';
-import { shouldFilterException } from 'src/engine/utils/global-exception-handler.util';
+import { shouldCaptureException } from 'src/engine/utils/global-exception-handler.util';
 import 'src/instrument';
 import { QueueWorkerModule } from 'src/queue-worker/queue-worker.module';
 
@@ -23,7 +23,7 @@ async function bootstrap() {
   } catch (err) {
     loggerService?.error(err?.message, err?.name);
 
-    if (!shouldFilterException(err)) {
+    if (shouldCaptureException(err)) {
       exceptionHandlerService?.captureExceptions([err]);
     }
